requirements.txt
amqp==5.3.1
annotated-types==0.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
boto3==1.35.76
botocore==1.35.76
Brotli==1.1.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
dj-database-url==2.3.0
dj-email-url==1.0.6
Django==5.1.4
django-allauth==65.3.0
django-allauth-ui==1.5.3
django-auth-style==0.8.1
django-celery-results==2.5.1
django-click==2.4.0
django-composed-configuration==0.25.0
django-configurations==2.5.1
django-cors-headers==4.6.0
django-extensions==3.2.3
django-filter==24.3
django-girder-utils==0.13.1
django-login-required-middleware==0.9.0
django-ninja==1.3.0
django-oauth-toolkit==3.0.1
django-s3-file-field==1.0.1
django-storages==1.14.4
django-vite==3.0.5
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
drf-yasg==1.21.8
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jmespath==1.0.1
jwcrypto==1.5.6
kombu==5.4.2
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.1.0
numpy==2.2.0
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg==3.2.3
psycopg-c==3.2.3
psycopg-pool==3.2.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.10.1
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
s3transfer==0.10.4
scipy==1.14.1
sentry-sdk==2.19.2
service-identity==24.2.0
six==1.17.0
slippers==0.6.2
sqlparse==0.5.2
Twisted==24.11.0
txaio==23.1.1
typeguard==2.13.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
-e .
zope.interface==7.2
